+++ src/read-dir-deep.ts
@@ -1,38 +1,9 @@
-import fs from 'fs';
 import globby from 'globby';
 import pathSort from 'path-sort2';
 
-function readDir(pathname: string) {
-    return new Promise((resolve, reject) => {
-        fs.readdir(pathname, (error, files) => {
-            if (error) {
-                reject(error);
-
-                return;
-            }
-
-            resolve(files);
-        });
-    });
-}
-
-function stat(pathname: string) {
-    return new Promise((resolve, reject) => {
-        fs.stat(pathname, (error, stats) => {
-            if (error) {
-                reject(error);
-
-                return;
-            }
-
-            resolve(stats);
-        });
-    });
-}
-
 type Options = { patterns?: string[] };
 
-const defaultPatterns = ['.'];
+const defaultPatterns = ['**'];
 const defaultOptions = {
     deep: true,
     dot: true,
@@ -41,14 +12,6 @@ const defaultOptions = {
 
 async function readDirDeep(startPath: string, options: Options = {}) {
     const { patterns = defaultPatterns, ...globbyOptions } = options;
-
-    const fileStats = await stat(startPath);
-    // @ts-ignore
-    const isDirectory = fileStats.isDirectory();
-    if (isDirectory === false) {
-        // will throw ENOTDIR
-        await readDir(startPath);
-    }
 
     const fileList = await globby(patterns, {
         cwd: startPath,
@@ -63,12 +26,6 @@ async function readDirDeep(startPath: string, options: Options = {}) {
 
 function readDirDeepSync(startPath: string, options: Options = {}) {
     const { patterns = defaultPatterns, ...globbyOptions } = options;
-
-    const isDirectory = fs.statSync(startPath).isDirectory();
-    if (isDirectory === false) {
-        // will throw ENOTDIR
-        fs.readdirSync(startPath);
-    }
 
     const fileList = globby.sync(patterns, {
         cwd: startPath,