+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+-   Change default pattern to `**`. Correctly throw `ENOTDIR` fix
+
 ## [4.0.0] - 2019-01-15
 
 -   Migrate to Typescript